Fixed some PR comments.
