+++ aloha-io-avro/src/main/scala/com/eharmony/aloha/semantics/compiled/plugin/avro/AvroSchema.scala
